refactor: remove `Suspense` in page
