+++ .changeset/spotty-coats-learn.md
@@ -1,0 +1,5 @@
+---
+"copipe": patch
+---
+
+Removed `Suspense`.
