+++ package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@mui/icons-material": "^6.0.0",
     "@next/third-parties": "15.1.6",
-    "@prisma/client": "^6.0.0",
+    "@prisma/client": "^6.3.1",
     "@supabase/supabase-js": "^2.39.2",
     "@types/react": "npm:types-react@19.0.0-rc.1",
     "@types/react-dom": "npm:types-react-dom@19.0.0",
