+++ src/app/(side-menu-layout)/archives/[id]/loading.tsx
@@ -2,7 +2,7 @@ import LoadingComment from "@/modules/comment/loading"
 import { CopipeCardItemSkelton } from "@/modules/copipeCardItemSkeleton"
 import { Container, VStack } from "@yamada-ui/react"
 
-export default function ArchivePageLoading() {
+export default function Loading() {
   return (
     <VStack>
       <Container>
