+++ src/app/(side-menu-layout)/page.tsx
@@ -1,7 +1,6 @@
 import { getHomePageCopipe } from "@/db/server/copipes"
 import { CopipeCardItem } from "@/modules/copipeCardItem"
 import CopipePagination from "@/modules/copipePagination"
-import SideMenuLayout from "@/modules/layouts/sideMenuLayout"
 import SearchForm from "@/modules/searchForm"
 import { Container, VStack } from "@yamada-ui/react"
 
@@ -9,25 +8,19 @@ export default async function Home() {
   const { copipes, count } = await getHomePageCopipe()
 
   return (
-    <SideMenuLayout>
-      <VStack>
-        <SearchForm />
+    <VStack>
+      <SearchForm />
 
-        <Container>
-          {copipes.map((copipe) => (
-            <CopipeCardItem
-              copipeItem={copipe}
-              key={`copipe-card-item-${copipe.id}`}
-            />
-          ))}
-        </Container>
+      <Container>
+        {copipes.map((copipe) => (
+          <CopipeCardItem
+            copipeItem={copipe}
+            key={`copipe-card-item-${copipe.id}`}
+          />
+        ))}
+      </Container>
 
-        <CopipePagination
-          page={1}
-          total={Math.ceil(count / 10)}
-          url="/search"
-        />
-      </VStack>
-    </SideMenuLayout>
+      <CopipePagination page={1} total={Math.ceil(count / 10)} url="/search" />
+    </VStack>
   )
 }
