+++ src/app/(side-menu-layout)/search/page.tsx
@@ -1,21 +1,49 @@
+import { fetchSearchCopipes } from "@/db/server/copipes"
+import { CopipeCardItem } from "@/modules/copipeCardItem"
+import CopipePagination from "@/modules/copipePagination"
+import SearchForm from "@/modules/searchForm"
+import { Container, Text, VStack } from "@yamada-ui/react"
 import { Metadata } from "next/types"
-import { Suspense } from "react"
-
-import { SearchPageLoading } from "./page-loading"
-import { SearchPageTemplate } from "./page-template"
 
 export const metadata: Metadata = {
   title: "検索",
 }
 
-export default function page({
-  searchParams,
+export default async function page({
+  searchParams: searchParamsProps,
 }: {
   searchParams: Promise<Record<string, string | string[] | undefined>>
 }) {
+  const searchParams = await searchParamsProps
+  const searchText =
+    typeof searchParams.text === "string" ? searchParams.text : ""
+  const page =
+    typeof searchParams.page === "string" ? Number(searchParams.page) : 1
+
+  const [copipes, count] = await fetchSearchCopipes(searchText, page)
+
   return (
-    <Suspense fallback={<SearchPageLoading />}>
-      <SearchPageTemplate searchParams={searchParams} />
-    </Suspense>
+    <VStack>
+      <SearchForm />
+
+      <Container>
+        {copipes.length == 0 ? (
+          <Text textStyle="noHit" variant="body1">
+            該当なし
+          </Text>
+        ) : (
+          copipes.map((copipe) => (
+            <CopipeCardItem copipeItem={copipe} key={copipe.id} />
+          ))
+        )}
+      </Container>
+
+      <CopipePagination
+        page={page}
+        params={{ name: "text", param: searchText }}
+        total={Math.ceil(count / 10)}
+        url="/search"
+      />
+    </VStack>
   )
 }
