+++ src/modules/loadingSearchForm.tsx
@@ -18,14 +18,11 @@ export default function LoadingSearchForm(props: { text?: string }) {
           <InputLeftElement>
             <Loading color="secondary" variant="dots" />
           </InputLeftElement>
-          <Input
-            focusBorderColor="secondary"
-            id="text"
-            isReadOnly
-            value={text}
-          />
+
+          <Input focusBorderColor="secondary" id="text" readOnly value={text} />
         </InputGroup>
       </Card>
+
       <Button apply="searchFormStyle.button" disabled variant="link">
         追加はこちらから
       </Button>
